Fixed jumping effect in case minHeight/minWidth is reached
When you resize the element with upper/left anchors and reach either min height
or min width the element sort of "jumps" up/left.

It happens because element's size is increased exceptionally setting it to minWidth/minHeight
while calculated newWidth/newHeight was smaller. The bad thing is that newX/newY are not
recalculated in this case which leads to jumping effect.

The fix is decreasing newX/newY with same amount as newWidth/newHeight are increased with thus
leaving the element at the same place.